Move GameClient to root directory, Remove unused import statement
