+++ src/com/plantpocalypse/controller/GameDirector.java
@@ -4,7 +4,6 @@ import com.plantpocalypse.model.Game;
 import com.plantpocalypse.util.Dialogue;
 import com.plantpocalypse.model.Player;
 import com.plantpocalypse.model.Room;
-import com.plantpocalypse.view.GameGUI;
 
 import java.util.HashMap;
 import java.util.List;
